FF-91 : move classes to their own folder, and params for each calc to one file
