+++ calculations/summer_baseflow.py
@@ -4,7 +4,7 @@ import numpy as np
 import pandas as pd
 from utils.helpers import is_multiple_date_data, smart_plot, remove_offset_from_julian_date
 from utils.matrix_convert import convert_raw_data_to_matrix, sort_matrix, insert_column_header
-from utils.calc_annual_flow_metrics import Gauge
+from classes.Gauge import Gauge
 
 np.warnings.filterwarnings('ignore')
