+++ classes/FlowExceedance.py
@@ -1,0 +1,11 @@
+class FlowExceedance:
+
+    def __init__(self, start_date, end_date, duration, exceedance):
+        self.start_date = start_date
+        self.end_date = end_date
+        self.duration = duration
+        self.flow = []
+        self.exceedance = exceedance
+
+    def add_flow(self, flow_data):
+        self.flow.append(flow_data)
